pyproject.toml
[project]
name = "rishta-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "syedmaazhasan0", email = "160649717+syedmaazhasan0@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "openai-agents>=0.2.2",
]

[project.scripts]
rishta-agent = "rishta_agent:main"
main = "rishta_agent.main:start"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
